Drop yaml conversion; add as_dict() to convert core Python types
Original code supported direct conversion (dump) to YAML,
but had issues in case of JSON, a similar dumper registration was misisng.
So this part is removed completely.

To solve the above difference, an as_dict() member is added to the Node
and a global wrapper, too, just like attrs.as_dict().